python2c++_frontier/requirements.txt
annotated-types==0.6.0
anthropic==0.26.0
anyio==3.7.1
beautifulsoup4==4.12.3
blinker==1.6.2
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.3
fsspec==2024.6.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
huggingface_hub==0.24.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.4.1
MarkupSafe==2.1.3
openai==1.37.0
packaging==24.1
pip==24.2
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.0
soupsieve==2.5
tokenizers==0.19.1
tqdm==4.66.5
typing_extensions==4.11.0
urllib3==2.2.3
Werkzeug==3.0.3
wheel==0.44.0
